tests/page_object/base_page_object.py
from appium.webdriver.common.mobileby import MobileBy
from selenium.webdriver.support import expected_conditions as ec
from selenium.webdriver.support.wait import WebDriverWait


class BasePage:

    def __init__(self, driver):
        self.driver = driver
        self.wait = WebDriverWait(self.driver, 10)

    def find_element(self, selector):
        element = self.wait.until(ec.visibility_of_element_located((MobileBy.ID, selector)))
        return element
