requirements.txt
python 3.6
pip 20.1
setuptools 46.1.3
pytest 5.4.1
Appium-Python-Client 0.52
